requirements.txt
fastapi
uvicorn
SQLAlchemy
asyncpg
pydantic-settings
python-multipart
a2wsgi
gunicorn
bcrypt==3.2.0
passlib
python-jose[cryptography]
aiosqlite
email-validator
firebase-admin

